% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotAnnotation.R
\name{AddShadingPlot}
\alias{AddShadingPlot}
\title{Add Shading to Various Summary Plots}
\usage{
AddShadingPlot(
  output_plot,
  x_start,
  x_end,
  y_start = NULL,
  y_end = NULL,
  col,
  alpha = 0.4
)
}
\arguments{
\item{output_plot}{The plot onto which you wish to add text}

\item{x_start, x_end}{The values on the x-axis (calendar age) at which you want the shading to start and end}

\item{y_start, y_end}{\code{NULL} if you want the shading to cover the entire y-axis range. Otherwise (if you want to plot a box) the y-axis values where you want the sahding to start and end}

\item{col}{The color to be used}

\item{alpha}{The level of transparency for the shading (\verb{0 < alpha < 1}) to see the plot underneath. \code{0} means fully transparent and \code{1} means fully opaque.}
}
\value{
None
}
\description{
Having plotted the output of the various modelling approaches, add a shaded interval or box to the plot. This function/annotation can be applied after having called any of \link{PlotPosteriorMean}
}
\examples{
# NOTE: This example is shown with a small n_iter and n_posterior_samples
# to speed up execution.
# Try n_iter and n_posterior_samples as the function defaults.

pp_output <- PPcalibrate(
    pp_uniform_phase$c14_age,
    pp_uniform_phase$c14_sig,
    intcal20,
    n_iter = 1000,
    show_progress = FALSE)

# Default plot with 2 sigma interval
posterior_mean_plot <- PlotPosteriorMeanRate(
    pp_output,
    n_posterior_samples = 100)

# Add transparent red shaded period to plot
AddShadingPlot(posterior_mean_plot,
    x_start = 620, x_end = 600,
    col = "red")

# Add green (more opaque) shaded box to plot
AddShadingPlot(posterior_mean_plot,
    x_start = 590, x_end = 550,
    y_start = 500, y_end = 410,
    col = "green", alpha = 0.8)

}
