% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotAnnotation.R
\name{AddTextPlot}
\alias{AddTextPlot}
\title{Add Text Annotation to Various Summary Plots}
\usage{
AddTextPlot(
  output_plot,
  x,
  y,
  labels,
  adj = NULL,
  pos = NULL,
  offset = 0.5,
  vfont = NULL,
  cex = 1,
  col = NULL,
  font = NULL
)
}
\arguments{
\item{output_plot}{The plot onto which you wish to add text}

\item{x, y}{numeric vectors of coordinates where the text
    \code{labels} should be written.  If the length of \code{x} and
    \code{y} differs, the shorter one is recycled.}

\item{labels}{a character vector or \link{expression} specifying
    the \emph{text} to be written.  An attempt is made to coerce other
    language objects (names and calls) to expressions, and vectors and
    other classed objects to character vectors by \code{\link{as.character}}.
    If \code{labels} is longer than \code{x} and
    \code{y}, \code{labels} is truncated to \code{max(length(x), length(y))}.}

\item{adj}{one or two values in \eqn{[0, 1]} which specify the x
    (and optionally y) adjustment (\sQuote{justification}) of the
    labels, with 0 for left/bottom, 1 for right/top, and 0.5 for
    centered.  On most devices values outside \eqn{[0, 1]} will also
    work.  See below.}

\item{pos}{a position specifier for the text.  If specified this
    overrides any \code{adj} value given.  Values of \code{1},
    \code{2}, \code{3} and \code{4}, respectively indicate
    positions below, to the left of, above and to the right of
    the specified \code{(x,y)} coordinates.}

\item{offset}{when \code{pos} is specified, this value controls the
    distance (\sQuote{offset}) of the text label from the specified
    coordinate in fractions of a character width.}

\item{vfont}{\code{NULL} for the current font family, or a character
    vector of length 2 for \code{\link[grDevices]{Hershey}} vector
    fonts.  The first element of
    the vector selects a typeface and the second element selects a
    style.  Ignored if \code{labels} is an expression.}

\item{cex}{numeric \bold{c}haracter \bold{ex}pansion factor; multiplied
    by \code{\link[graphics]{par}("cex")} yields the final character size.
    \code{NULL} and \code{NA} are equivalent to \code{1.0}.}

\item{col, font}{the color and (if \code{vfont = NULL}) font to be
    used, possibly vectors.  These default to the values of the global
    \link[graphics]{graphical parameters} in \code{\link[graphics]{par}()}.}
}
\value{
None
}
\description{
Having plotted the output of the various modelling approaches, add text annotation at a given location on the plot. It is based on the base \link{text} function

This function/annotation can be applied after having called any of \link{PlotPosteriorMeanRate}, \link{PlotPosteriorChangePoints}, \link{CalibrateSingleDetermination}, \link{PlotPredictiveCalendarAgeDensity}, \link{PlotCalendarAgeDensityIndividualSample}, or \link{PlotRateIndividualRealisation}.
}
\examples{
# NOTE: This example is shown with a small n_iter and n_posterior_samples
# to speed up execution.
# Try n_iter and n_posterior_samples as the function defaults.

pp_output <- PPcalibrate(
    pp_uniform_phase$c14_age,
    pp_uniform_phase$c14_sig,
    intcal20,
    n_iter = 1000,
    show_progress = FALSE)

# Default plot with 2 sigma interval
posterior_mean_plot <- PlotPosteriorMeanRate(
    pp_output,
    n_posterior_samples = 100)

# Add horizontal green dashed line
AddLinePlot(
     posterior_mean_plot,
     h = 500,
     col = "darkgreen",
     lwd = 1,
     lty = 2)

# Add text to annotate line
AddTextPlot(posterior_mean_plot,
    x = 575, y = 500,
    labels = expression(paste("500", " "^14, "C ", "yrs BP")),
    cex = 0.7,
    pos = 3,
    offset = 0.2,
    col = "darkgreen")

}
