% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{armit}
\alias{armit}
\title{Example real-life data - Population Decline in Iron Age Ireland}
\format{
\subsection{\code{armit}}{

A data frame with 2021 rows and 4 columns:
\describe{
\item{c14_age}{The 14C age (yr BP)}
\item{c14_sig}{The uncertainty in the 14C age reported by the radiocarbon laboratory}
\item{f14c_age}{The F14C concentration}
\item{f14c_sig}{The uncertainty in the F14C concentration reported by the radiocarbon laboratory}
}
}
}
\source{
\url{http://doi.org/10.1073/pnas.1408028111}
}
\usage{
armit
}
\description{
2021 radiocarbon determinations collated by Armit et al. from archaeological
groups operating in Ireland, to investigate whether a wetter environment
around 2700 cal yr BP led to a population collapse. Ref:
Armit, I., Swindles, G.T., Becker, K., Plunkett, G., Blaauw, M., 2014. Rapid
climate change did not cause population collapse at the end of the European
Bronze Age. Proceedings of the National Academy of Sciences 111, 17045–17049.
}
\keyword{datasets}
