% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BivarGibbsDirichletwithSlice.R
\name{BivarGibbsDirichletwithSlice}
\alias{BivarGibbsDirichletwithSlice}
\title{Performs the Gibbs sampler to estimate the density of
a set of observations using a Dirichlet Mixture Density}
\usage{
BivarGibbsDirichletwithSlice(
  x,
  xsig,
  lambda,
  nu1,
  nu2,
  A,
  B,
  mualpha,
  sigalpha,
  alphaprshape,
  alphaprrate,
  niter = 100,
  nthin = 10,
  theta = NA,
  alphapr = "gamma",
  w = 200,
  m = 50,
  calibration_curve,
  nclusinit = 10,
  sensibleinit = TRUE,
  showprogress = TRUE
)
}
\arguments{
\item{x}{TODO}

\item{xsig}{TODO}

\item{lambda, nu1, nu2}{Hyperparameters for the prior on the means
\eqn{\phi_j} and precision \eqn{\tau_j} of each individual calendar age
cluster \eqn{j}.
\deqn{(\phi_j, \tau_j)|\mu_{\phi} \sim
\textrm{NormalGamma}(\mu_{\phi}, \lambda, \nu_1, \nu_2)} where
\eqn{\mu_{\phi}} is the overall cluster centering.}

\item{A, B}{Prior on \eqn{\mu_{\phi}} giving the mean and precision of the
overall centering \eqn{\mu_{\phi} \sim N(A, B^{-1})} i.e.
B small is uninformative.}

\item{mualpha}{TODO}

\item{sigalpha}{TODO}

\item{alphaprshape}{TODO}

\item{alphaprrate}{TODO}

\item{niter}{The number of MCMC iterations (optional). Default is 100.}

\item{nthin}{How much to thin the output (optional). 1 is no thinning,
a larger number is more thinning. Default is 10. Must choose an integer more
than 1 and not too close to \code{n_iter}, since after burn-in there are
\eqn{(n_{\textrm{iter}}/n_{\textrm{thin}})/2} samples from posterior to
potentially use.}

\item{theta}{The initial estimate for the underlying calendar ages
(optional). If supplied it must be a vector with the same length as
\code{c14_determinations}. Will be overridden if \code{sensible_initialisation} is
\code{TRUE}.}

\item{alphapr}{TODO}

\item{w}{TODO}

\item{m}{TODO}

\item{calibration_curve}{TODO}

\item{nclusinit}{TODO}

\item{sensibleinit}{TODO}

\item{showprogress}{Whether to show a progress bar in the console during
execution. Default is \code{TRUE}.}
}
\value{
A list containing 6 items
}
\description{
THIS CONSIDERS BOTH MEAN AND VARIANCE OF THE CLUSTERS TO BE UNKNOWN
}
