% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/independent_calibration.R
\name{CalibrateSingleDetermination}
\alias{CalibrateSingleDetermination}
\title{Calibrates a single radiocarbon determination}
\usage{
CalibrateSingleDetermination(
  rc_determination,
  rc_uncertainty,
  calibration_data
)
}
\arguments{
\item{rc_determination}{A single observed radiocarbon determination (c14 age)}

\item{rc_uncertainty}{The uncertainty of the radiocarbon determination}

\item{calibration_data}{A dataframe which should contain one column entitled
c14_age and one column entitled c14_sig}
}
\value{
A vector containing the probability for each calendar age in the
calibration data
}
\description{
Calibrates a single radiocarbon determination
}
