% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/independent_calibration.R
\name{FindSPD}
\alias{FindSPD}
\title{Find the summed probability function (SPD) for a set of observations}
\usage{
FindSPD(
  calendar_age_range,
  c14_determinations,
  c14_uncertainties,
  calibration_curve
)
}
\arguments{
\item{calendar_age_range}{An array of length 2 with the start and end
calendar age to calculate the SPD over}

\item{c14_determinations}{An array of observed radiocarbon determinations}

\item{c14_uncertainties}{An array of the radiocarbon determinations
uncertainties. Must be the same length as \code{c14_determinations}.}

\item{calibration_curve}{A dataframe which should contain at least 3 columns
entitled calendar_age, c14_age and c14_sig.
This format matches \link{intcal20}.}
}
\value{
A data frame with one column \code{calendar_age} containing the calendar
ages, and the other column \code{probability} containing the probability at that
calendar age
}
\description{
Takes a set of radiocarbon determinations and uncertainties and independently
calibrates each one, and then averages them to give the SPD estimate.
}
\examples{
FindSPD(
   calendar_age_range=c(600, 1600),
   c14_determinations=c(602, 805, 1554),
   c14_uncertainties=c(35, 34, 45),
   calibration_curve=intcal20)
}
