% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{InterpolateCalibrationCurve}
\alias{InterpolateCalibrationCurve}
\title{Interpolate calendar ages for a calibration curve}
\usage{
InterpolateCalibrationCurve(new_calendar_ages, calibration_curve)
}
\arguments{
\item{new_calendar_ages}{A scalar or vector containing calendar ages to
interpolate the calibration curve to}

\item{calibration_curve}{A dataframe which should contain at least 3 columns
entitled calendar_age, c14_age and c14_sig.
This format matches \link{intcal20}.}
}
\value{
A new dataframe with entries for the interpolated \code{c14_age} and
\code{c14_sig} values at the \code{calendar_age} values given in \code{new_calendar_ages}
}
\description{
Interpolate calendar ages for a calibration curve
}
\examples{
InterpolateCalibrationCurve(51020.5, intcal20)
InterpolateCalibrationCurve(c(51020.5, 51021.5), intcal20)
}
