% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kerr}
\alias{kerr}
\title{Example real-life data - Irish Rath}
\format{
\subsection{\code{kerr}}{

A data frame with 255 rows and 4 columns:
\describe{
\item{c14_age}{The 14C age (yr BP)}
\item{c14_sig}{The uncertainty in the 14C age reported by the radiocarbon laboratory}
\item{f14c}{The F14C concentration}
\item{f14c_sig}{The uncertainty in the F14C concentration reported by the radiocarbon laboratory}
}
}
}
\source{
\url{http://doi.org/10.1016/j.jas.2013.09.002}
}
\usage{
kerr
}
\description{
255 radiocarbon determinations collated by Kerr and McCormick related to the
building and use of raths in Ireland in the early-medieval period. Ref:
Kerr, T., McCormick, F., 2014. Statistics, sunspots and settlement:
influences on sum of probability curves.
Journal of Archaeological Science 41, 493 – 501.
}
\keyword{datasets}
