% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{marine04}
\alias{marine04}
\title{Marine04 calibration curve}
\format{
\subsection{\code{marine04}}{

A data frame with 3,301 rows and 5 columns providing the Marine04 radiocarbon age
calibration curve on a calendar grid spanning from 50,000--0 cal yr BP:
\describe{
\item{calendar_age}{The calendar age (in cal yr BP)}
\item{c14_age}{The \eqn{{}^{14}}C age (in \eqn{{}^{14}}C yr BP)}
\item{c14_sig}{The (1\eqn{\sigma}) uncertainty in the \eqn{{}^{14}}C age}
\item{f14c}{The \eqn{{}^{14}}C age expressed as F\eqn{{}^{14}}C concentration}
\item{f14c_sig}{The (1\eqn{\sigma}) uncertainty in the F\eqn{{}^{14}}C concentration}
}
}
}
\source{
http://doi.org/10.1017/s0033822200033002
}
\usage{
marine04
}
\description{
The Marine04 marine radiocarbon age calibration curve
on a calendar grid spanning from 50,000--0 cal yr BP
(Before Present, 0 cal yr BP corresponds to 1950 CE). \cr \cr
\emph{Note:} This dataset provides \eqn{{}^{14}}C ages and F\eqn{{}^{14}}C values
on a calendar age grid. This is a different format from oxcal/calib .14c files
which give the \eqn{{}^{14}}C ages and \eqn{{\Delta}^{14}}C values.\cr \cr
\strong{Reference:} \cr Hughen KA, Baillie MGL, Bard E, Beck JW,
Bertrand CJH, Blackwell PG, Buck CE, Burr GS, Cutler KB, Damon PE,
Edwards RL, Fairbanks RG, Friedrich M, Guilderson TP, Kromer B,
McCormac G, Manning S, Bronk Ramsey C, Reimer PJ, Reimer RW,
Remmele S, Southon JR, Stuiver M, Talamo S, Taylor FW,
van der Plicht J, Weyhenmeyer CE. 2004. Marine04 marine radiocarbon age calibration, 0-26 cal kyr BP.
\emph{Radiocarbon} \strong{46}(3):1059-1086 https://doi.org/10.1017/s0033822200033002. \cr \cr
}
\keyword{datasets}
