% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{marine20}
\alias{marine20}
\title{Marine20 calibration curve}
\format{
\subsection{\code{marine20}}{

A data frame with 5,501 rows and 5 columns providing the Marine20 radiocarbon age
calibration curve on a calendar grid spanning from 55,000--0 cal yr BP:
\describe{
\item{calendar_age}{The calendar age (in cal yr BP)}
\item{c14_age}{The \eqn{{}^{14}}C age (in \eqn{{}^{14}}C yr BP)}
\item{c14_sig}{The (1\eqn{\sigma}) uncertainty in the \eqn{{}^{14}}C age}
\item{f14c}{The \eqn{{}^{14}}C age expressed as F\eqn{{}^{14}}C concentration}
\item{f14c_sig}{The (1\eqn{\sigma}) uncertainty in the F\eqn{{}^{14}}C concentration}
}
}
}
\source{
http://doi.org/10.1017/RDC.2020.68
}
\usage{
marine20
}
\description{
The Marine20 marine radiocarbon age calibration curve
on a calendar grid spanning from 55,000--0 cal yr BP
(Before Present, 0 cal yr BP corresponds to 1950 CE). \cr \cr
\emph{Note:} This dataset provides \eqn{{}^{14}}C ages and F\eqn{{}^{14}}C values
on a calendar age grid. This is a different format from oxcal/calib .14c files
which give the \eqn{{}^{14}}C ages and \eqn{{\Delta}^{14}}C values.\cr \cr
\strong{Reference:} \cr Heaton TJ, Köhler P, Butzin M, Bard E, Reimer RW,
Austin WEN, Bronk Ramsey C, Grootes PM, Hughen KA, Kromer B, Reimer PJ,
Adkins J, Burke A, Cook MA, Olsen J, Skinner L. 2020 Marine20 - the marine radiocarbon age calibration curve (0–55,000 cal BP).
\emph{Radiocarbon} \strong{62} https://doi.org/10.1017/RDC.2020.68. \cr \cr
}
\keyword{datasets}
