% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_calendar_age_density.R
\name{PlotCalendarAgeDensity}
\alias{PlotCalendarAgeDensity}
\title{Plots the predicted calendar age density from the output data}
\usage{
PlotCalendarAgeDensity(
  c14_determinations,
  c14_uncertainties,
  calibration_curve,
  output_data,
  n_posterior_samples,
  show_SPD = TRUE,
  xlimscal = 1,
  ylimscal = 1,
  denscale = 3
)
}
\arguments{
\item{c14_determinations}{A vector containing the radiocarbon determinations.}

\item{c14_uncertainties}{A vector containing the radiocarbon determination
uncertainties. Must be the same length as \code{c14_determinations}.}

\item{calibration_curve}{A dataframe which should contain one column entitled
c14_age and one column entitled c14_sig.
This format matches \link{intcal20}.}

\item{output_data}{The return value from one of the updating functions e.g.
\link{WalkerBivarDirichlet} or
\link{BivarGibbsDirichletwithSlice}.}

\item{n_posterior_samples}{Current number of samples it will draw from this
posterior to estimate the calendar age density (possibly repeats).}

\item{show_SPD}{Whether to calculate and show the summed probability function
on the plot (optional). Default is \code{TRUE}.}

\item{xlimscal, ylimscal}{Whether to scale the x or y limits (optional).
Default is 1. Values more than 1 will increase range of the limits,
values less than 1 will decrease the range of the limits.}

\item{denscale}{Whether to scale the vertical range of the density plot
relative to the calibration curve plot (optional). Default is 3 which means
that the maximum SPD density will be at 1/3 of the height of the plot.
TODO: base on the output data.}
}
\value{
A dataframe containing the \code{calendar_age}, the \code{density_mean}
and the 95\% confidence intervals for the density \code{density_ci_lower} and
\code{density_ci_upper}.
}
\description{
Plots the input radiocarbon determinations and calibration curve, with the
output predicted density on the same plot. Can also optionally show the
SPD estimate.
}
