% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPredictiveCalendarAgeDensity.R
\name{PlotPredictiveCalendarAgeDensity}
\alias{PlotPredictiveCalendarAgeDensity}
\title{Plots the calendar age density of all objects from the output data}
\usage{
PlotPredictiveCalendarAgeDensity(
  output_data,
  n_posterior_samples,
  calibration_curve = NULL,
  show_SPD = FALSE,
  show_confidence_intervals = TRUE,
  interval_width = "2sigma",
  bespoke_probability = NA,
  true_density = NULL,
  xlimscal = 1,
  ylimscal = 1,
  denscale = 3,
  n_calc = 1001
)
}
\arguments{
\item{output_data}{The return value from one of the updating functions e.g.
\link{WalkerBivarDirichlet} or
\link{PolyaUrnBivarDirichlet} or a list, each item containing
one of these values. Optionally, the output data can have an extra list item
named \code{label} which is used to set the label on the plot legend.}

\item{n_posterior_samples}{Current number of samples it will draw from this
posterior to estimate the calendar age density (possibly repeats).}

\item{calibration_curve}{The name of the calibration curve variable is saved
in the output data. However if the variable with this name is no longer in
your environment then you should pass the calibration curve here. If
provided this should be a dataframe which should contain at least 3 columns
entitled calendar_age, c14_age and c14_sig.
This format matches \link{intcal20}.}

\item{show_SPD}{Whether to calculate and show the summed probability function
on the plot (optional). Default is \code{FALSE}.}

\item{show_confidence_intervals}{Whether to show the confidence intervals
for the chosen probability on the plot. Default is \code{TRUE}.}

\item{interval_width}{The confidence intervals to show for both the
calibration curve and the predictive density. Choose from one of "1sigma",
"2sigma" and "bespoke". Default is "2sigma".}

\item{bespoke_probability}{The probability to use for the confidence interval
if "bespoke" is chosen above. E.g. if 0.95 is chosen, then the 95\% confidence
interval is calculated. Ignored if "bespoke" is not chosen.}

\item{true_density}{The true calendar age density (optional). Expects a data
frame with two columns, the first column being calendar age and the right
column being the normalized density.}

\item{xlimscal, ylimscal}{Whether to scale the x or y limits (optional).
Default is 1. Values more than 1 will increase range of the limits,
values less than 1 will decrease the range of the limits.}

\item{denscale}{Whether to scale the vertical range of the density plot
relative to the calibration curve plot (optional). Default is 3 which means
that the maximum SPD density will be at 1/3 of the height of the plot.}

\item{n_calc}{Number of points to use when calculating the predictive
density. Default is 1001.}
}
\value{
A list, each item containing a data frame of the \code{calendar_age}, the
\code{density_mean} and the 95\% confidence intervals for the density
\code{density_ci_lower} and \code{density_ci_upper} for each set of output data.
}
\description{
Plots the input radiocarbon determinations and calibration curve, with the
output predicted density on the same plot. Can also optionally show the
SPD estimate
}
\examples{
# Plot results for a single calibration
PlotPredictiveCalendarAgeDensity(walker_example_output, 500)

# Plot results from a calibration, and add a label
new_output = walker_example_output
new_output$label = "My plot"
PlotPredictiveCalendarAgeDensity(new_output, 500)

# Plot results from two calibrations on the same plot, and show the SPD
PlotPredictiveCalendarAgeDensity(
  list(walker_example_output, polya_urn_example_output), 500, show_SPD = TRUE)

# Plot and show the 1-sigma confidence interval
PlotPredictiveCalendarAgeDensity(walker_example_output, 500, interval_width = "1sigma")

# Plot and show the 80\% confidence interval
PlotPredictiveCalendarAgeDensity(
  walker_example_output, 500, interval_width = "bespoke", bespoke_probability = 0.8)
}
