% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pp_uniform_phase_marine}
\alias{pp_uniform_phase_marine}
\title{Example artificial marine data - Uniform Phase}
\format{
\subsection{\code{pp_uniform_phase_marine}}{

A data frame with 40 rows and 7 columns:
\describe{
\item{c14_age}{The simulated \eqn{{}^{14}}C age (in \eqn{{}^{14}}C yr BP)}
\item{c14_sig}{The (fixed) \eqn{{}^{14}}C age measurement uncertainty used in the simulation (set at 15 \eqn{{}^{14}}C yrs)}
\item{sample_source}{A character vector to clarify that the data were simultaed using the Marine20 calibration curve}
\item{delta_r}{The \eqn{\Delta R} offset, set at 100 \eqn{{}^{14}}C yrs for marine samples}
\item{delta_r_sig}{The (1\eqn{\sigma}) uncertainty in the \eqn{\Delta R} offset. Set at 20 \eqn{{}^{14}}C yrs}
\item{f14c}{The corresponding simulated values of F\eqn{{}^{14}}C concentration}
\item{f14c_sig}{The (fixed) corresponding F\eqn{{}^{14}}C measurement uncertainty used in the simulation}
}
}
}
\usage{
pp_uniform_phase_marine
}
\description{
40 simulated marine radiocarbon determinations that are created as though they were marine samples (with a \eqn{\Delta R} offset of 100 \eqn{\pm} 20 \eqn{{}^{14}}C yr). As for \link{pp_uniform_phase}, the samples are simulated with underlying calendar ages drawn (uniformly at random) from the period  550--500 cal yr BP
\deqn{f(\theta) = U[550, 500]}
The corresponding \eqn{{}^{14}}C ages are then simulated based upon the Marine20 calibration curve with a \eqn{\Delta R} of 100 \eqn{\pm} 20 \eqn{{}^{14}}C yr. The observational uncertainty of each determination is set to be 15 \eqn{{}^{14}}C yrs. \cr \cr
The corresponding \eqn{{}^{14}}C ages are then simulated based upon the Marine20 calibration curve
(convolved with the 15 \eqn{{}^{14}}C yr measurement uncertainty):
\deqn{X_i | \theta_i \sim N(m(\theta_i), \rho(\theta_i)^2 + 15^2),}
where \eqn{m(\theta_i)} and \eqn{\rho(\theta_i)} are the calibration curve pointwise
means and uncertainties. The simulated \eqn{\Delta R} of 100 \eqn{\pm} 20 \eqn{{}^{14}}C yr is then added.  \cr \cr
}
\keyword{datasets}
