% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shcal04}
\alias{shcal04}
\title{SHCal04 calibration curve}
\format{
\subsection{\code{shcal04}}{

A data frame with 2,202 rows and 5 columns:
\describe{
\item{calendar_age}{The calendar age (yr BP)}
\item{c14_age}{The \eqn{{}^{14}}C age (yr BP)}
\item{c14_sig}{The uncertainty in the \eqn{{}^{14}}C age reported by the radiocarbon laboratory}
\item{f14c}{The radiocarbon age expressed as F14C conctration}
\item{f14c_sig}{The uncertainty in F14C}
}
}
}
\source{
\url{http://doi.org/10.1017/S0033822200033014}
}
\usage{
shcal04
}
\description{
SHCal04 atmospheric curve (McCormac et al 2004)
McCormac et al. 2004
FG McCormac, AG Hogg, PG Blackwell, CE Buck, TFG Higham, and PJ Reimer
2004
SHCal04 Southern Hemisphere Calibration 0 - 11.0 cal kyr BP
Radiocarbon 46(3):1087-1092. doi:10.1017/S0033822200033014
}
\keyword{datasets}
