% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shcal20}
\alias{shcal20}
\title{SHCal20 calibration curve}
\format{
\subsection{\code{shcal20}}{

A data frame with 9,501 rows and 5 columns:
\describe{
\item{calendar_age}{The calendar age (yr BP)}
\item{c14_age}{The \eqn{{}^{14}}C age (yr BP)}
\item{c14_sig}{The uncertainty in the \eqn{{}^{14}}C age reported by the radiocarbon laboratory}
\item{f14c}{The radiocarbon age expressed as F14C conctration}
\item{f14c_sig}{The uncertainty in F14C}
}
}
}
\source{
\url{http://doi.org/10.1017/RDC.2020.59}
}
\usage{
shcal20
}
\description{
Atmospheric data from Hogg et al (2020)
Hogg et al. 2020
Hogg AG, Heaton TJ, Hua Q, Palmer JG, Turney CSM, Southon J, Bayliss A, Blackwell PG,
Boswijk G, Bronk Ramsey C, Pearson C, Petchey F, Reimer P, Reimer R, Wacker L.
2020.
SHCal20 Southern Hemisphere calibration, 0-55,000 years cal BP.
Radiocarbon 62. doi: 10.1017/RDC.2020.59
}
\keyword{datasets}
