% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateCalendarAgesGibbs.R
\name{UpdateCalendarAgesGibbs}
\alias{UpdateCalendarAgesGibbs}
\title{Title}
\usage{
UpdateCalendarAgesGibbs(
  likelihood_calendar_ages_from_calibration_curve,
  calendar_age_grid,
  rate_s,
  rate_h
)
}
\arguments{
\item{likelihood_calendar_ages_from_calibration_curve}{A matrix with column i being vector of
likelihood of theta (for each value in calendar_age_grid) of ith rc_determination}

\item{calendar_age_grid}{The possible set of calendar ages the observations can take}

\item{rate_s}{Current set of (calendar age) changepoints in rate of
piecewise constant Poisson process}

\item{rate_h}{Current set of heights/rates in each section of
piecewise constant Poisson process}
}
\value{
TODO
}
\description{
Title
}
\examples{
# TODO
}
