% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdatePoissonProcessRateRevJump.R
\name{UpdatePoissonProcessRateRevJump}
\alias{UpdatePoissonProcessRateRevJump}
\title{Title}
\usage{
UpdatePoissonProcessRateRevJump(
  theta,
  rate_s,
  rate_h,
  integrated_rate,
  prior_h_shape,
  prior_h_rate,
  prior_n_internal_changepoints_lambda,
  prob_move
)
}
\arguments{
\item{theta}{Vector of the observed times of the events}

\item{rate_s}{The current changepoints in the rate}

\item{rate_h}{The current heights in the rate (corresponding to above)}

\item{integrated_rate}{The integral of piecewise constant rate i.e.,
integral_0^L nu(t) dt}

\item{prior_h_shape, prior_h_rate}{prior parameters on heights
We assume each height in piecewise rate has h ~ Gamma(shape, rate)}

\item{prior_n_internal_changepoints_lambda}{Prior on number of internal changepoints n ~ Po(lambda)}

\item{prob_move}{Dataframe with probability of each type of RJ move
(change_pos, change_height, birth, death)}
}
\value{
TODO
}
\description{
Title
}
\examples{
# TO DO
}
