% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alces}
\alias{alces}
\title{Example real-life data - Alces in Yukon and Alaska}
\format{
\subsection{\code{alces}}{

A data frame with 58 rows and 7 columns:
\describe{
\item{lab_code}{The sample code for the \eqn{{}^{14}}C laboratory}
\item{site_code}{The site/museum code}
\item{location}{The location of the sample}
\item{c14_age}{The 14C age (yr BP)}
\item{c14_sig}{The uncertainty in the 14C age reported by the radiocarbon laboratory}
\item{f14c}{The F14C concentration}
\item{f14c_sig}{The uncertainty in the F14C concentration reported by the radiocarbon laboratory}
}
}
}
\source{
\url{https://doi.org/10.1038/nature04604}
}
\usage{
alces
}
\description{
58 radiocarbon determinations collated by Dale Guthrie, R. related to
Alces (moose) in Yukon and Alaska. Samples
are restricted to those between 25,000--6000 \eqn{{}^{14}}C yrs BP. \cr \cr
Ref: Dale Guthrie, R. New carbon dates link climatic change with human colonization
and Pleistocene extinctions. \emph{Nature} \strong{441}, 207--209 (2006).
https://doi.org/10.1038/nature04604
}
\keyword{datasets}
