% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{buchanan}
\alias{buchanan}
\title{Example real-life data - Palaeo-Indian demography}
\format{
\subsection{\code{buchanan}}{

A data frame with 628 rows and 4 columns:
\describe{
\item{c14_age}{The 14C age (yr BP)}
\item{c14_sig}{The uncertainty in the 14C age reported by the radiocarbon laboratory}
\item{f14c}{The F14C concentration}
\item{f14c_sig}{The uncertainty in the F14C concentration reported by the radiocarbon laboratory}
}
}
}
\source{
\url{http://doi.org/10.1073/pnas.0803762105}
}
\usage{
buchanan
}
\description{
628 radiocarbon determinations collated by Buchanan et al. representing the
ages of distinct archaeological  sites found across Canada and North America
during the time of the palaeoindians. Ref:
Buchanan, B., Collard, M., Edinborough, K., 2008. Paleoindian demography and
the extraterrestrial impact hypothesis. Proceedings of the National Academy
of Sciences 105, 11651–11654.
}
\keyword{datasets}
