% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C14AgeToOxcal.R
\name{C14AgeToOxcal}
\alias{C14AgeToOxcal}
\title{Outputs text suitable for running in OxCal}
\usage{
C14AgeToOxcal(
  model_name,
  c14_age,
  c14_sig,
  c14_name = NULL,
  outfile_path = NULL
)
}
\arguments{
\item{model_name}{The name given to the model in the OxCal code.}

\item{c14_age}{An array of 14C determinations in yr BP.}

\item{c14_sig}{An array of the errors for the 14C determinations, must be the same length as c14_age.}

\item{c14_name}{Optional. The name of each data point - if given it must be the same length of c14_age.}

\item{outfile_path}{Optional. If given the OxCal code will be output to the file at the path given, otherwise it
will be output to the terminal.}
}
\description{
Outputs text suitable for running in OxCal
}
\examples{
# Generate names automatically and outputs to the screen
C14AgeToOxcal("My_data", c(1123, 1128, 1135), c(32, 24, 25))

}
