% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalibrateSingleDetermination.R
\name{CalibrateSingleDetermination}
\alias{CalibrateSingleDetermination}
\title{Calibrate a single radiocarbon determination}
\usage{
CalibrateSingleDetermination(
  rc_determination,
  rc_sigma,
  calibration_curve,
  F14C_inputs = FALSE,
  plot_output = FALSE,
  interval_width = "2sigma",
  bespoke_probability = NA
)
}
\arguments{
\item{rc_determination}{A single observed radiocarbon determination
(\eqn{{}^{14}}C BP age or F14C concentration)}

\item{rc_sigma}{The uncertainty of the radiocarbon determination in the same units}

\item{calibration_curve}{A dataframe which must contain one column \code{calendar_age_BP}, and also
columns \code{c14_age} and \code{c14_sig} or \code{f14c} and \code{f14c_sig} (or both sets).
This format matches the curves supplied with this package e.g. \link{intcal20},
which contain all 5 columns.}

\item{F14C_inputs}{\code{TRUE} if the provided rc_determinations are F14C concentrations and \code{FALSE}
if they are radiocarbon age BP. Defaults to \code{FALSE}.}

\item{plot_output}{\code{TRUE} if you wish to plot the determination, calibration curve, and the
posterior calibrated age estimate. Defaults to \code{FALSE}}

\item{interval_width}{Only for usage when \code{plot_output == TRUE}. The confidence intervals to show for the
calibration curve and for the highest posterior density ranges.
Choose from one of "1sigma" (68.3\%), "2sigma" (95.4\%) and "bespoke". Default is "2sigma".}

\item{bespoke_probability}{The probability to use for the confidence interval
if "bespoke" is chosen above. E.g. if 0.95 is chosen, then the 95\% confidence
interval is calculated. Ignored if "bespoke" is not chosen.}
}
\value{
A data frame with one column \code{calendar_age_BP} containing the calendar
ages, and the other column \code{probability} containing the probability at that
calendar age.
}
\description{
Uses the supplied calibration curve to take a single radiocarbon
determination and uncertainty (expressed as the radiocarbon age BP or F14C concentration) and
calculate the calendar age probability density for it.
}
\examples{
# Calibration of a single determination expressed as 14C age BP
calib <- CalibrateSingleDetermination(860, 35, intcal20)
plot(calib, type = "l", xlim = c(1000, 600))

# Incorporating an automated plot to visualise the calibration
calib <- CalibrateSingleDetermination(860, 35, intcal20, plot_output = TRUE)

# Calibration of a single (old) determination expressed as 14C age BP
calib <- CalibrateSingleDetermination(
    31020, 100, intcal20)
plot(calib, type = "l", xlim = c(36500, 34500))

# Calibration of a single determination expressed as F14C concentration
calib <- CalibrateSingleDetermination(
    0.02103493, 0.0002618564, intcal20, F14C_inputs = TRUE)
plot(calib, type = "l", xlim = c(36500, 34500))

# Calibration of a single determination expressed as 14C age BP
# against SHCal20 (and creating an automated plot)
calib <- CalibrateSingleDetermination(1413, 25, shcal20, plot_output = TRUE)

# Implementing a bespoke confidence interval level
calib <- CalibrateSingleDetermination(1413, 25, shcal20,
    plot_output = TRUE, interval_width = "bespoke", bespoke_probability = 0.8)
}
