% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalibrateSingleDetermination.R
\name{CalibrateSingleDetermination}
\alias{CalibrateSingleDetermination}
\title{Calibrate a single radiocarbon determination}
\usage{
CalibrateSingleDetermination(c14_determination, c14_sigma, calibration_curve)
}
\arguments{
\item{c14_determination}{A single observed radiocarbon determination (c14 age)}

\item{c14_sigma}{The uncertainty of the radiocarbon determination}

\item{calibration_curve}{A dataframe which should contain at least 3 columns
entitled \code{calendar_age}, \code{c14_age} and \code{c14_sig}.
This format matches \link{intcal20}.}
}
\value{
A data frame with one column \code{calendar_age} containing the calendar
ages, and the other column \code{probability} containing the probability at that
calendar age.
}
\description{
Uses the supplied calibration curve to take a single radiocardbon
determination and uncertainty and calculate the calendar age probability
density for it.
}
\examples{
CalibrateSingleDetermination(31020, 35, intcal20)
}
