% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalibrateSingleDetermination.R
\name{CalibrateSingleDetermination}
\alias{CalibrateSingleDetermination}
\title{Calibrate a single radiocarbon determination}
\usage{
CalibrateSingleDetermination(
  rc_determination,
  rc_sigma,
  calibration_curve,
  F14C_inputs = FALSE
)
}
\arguments{
\item{rc_determination}{A single observed radiocarbon determination
(14C BP age or F14C concentration)}

\item{rc_sigma}{The uncertainty of the radiocarbon determination in the same units}

\item{calibration_curve}{A dataframe which must contain one column \code{calendar_age_BP}, and also
columns \code{c14_age} and \code{c14_sig} or \code{f14c} and \code{f14c_sig} (or both sets).
This format matches the curves supplied with this package e.g. \link{intcal20},
which contain all 5 columns.}

\item{F14C_inputs}{\code{TRUE} if the provided rc_determinations are F14C concentrations and \code{FALSE}
if they are radiocarbon age BP. Defaults to \code{FALSE}.}
}
\value{
A data frame with one column \code{calendar_age_BP} containing the calendar
ages, and the other column \code{probability} containing the probability at that
calendar age.
}
\description{
Uses the supplied calibration curve to take a single radiocarbon
determination and uncertainty (expressed as the radiocarbon age BP or F14C concentration) and
calculate the calendar age probability density for it.
}
\examples{
# Calibration of a single determination expressed as 14C age BP
calib = CalibrateSingleDetermination(31020, 35, intcal20)
plot(calib, type = "l", xlim = c(36000, 34000))

# Calibration of a single determination expressed as F14C concentration
calib = CalibrateSingleDetermination(
    0.02103493, 9.164975e-05, intcal20, F14C_inputs = TRUE)
plot(calib, type = "l", xlim = c(36000, 34000))
}
