% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalibrateSingleDetermination.R
\name{CalibrateSingleDetermination}
\alias{CalibrateSingleDetermination}
\title{Calibrate a single radiocarbon determination}
\usage{
CalibrateSingleDetermination(c14_determination, c14_sigma, calibration_curve)
}
\arguments{
\item{c14_determination}{A single observed radiocarbon determination (c14 age)}

\item{c14_sigma}{The uncertainty of the radiocarbon determination}

\item{calibration_curve}{A dataframe which should contain at least 3 columns
entitled calendar_age, c14_age and c14_sig.
This format matches \link{intcal20}.}
}
\value{
A vector containing the probability for each calendar age in the
calibration data
}
\description{
Uses the supplied calibration curve to take a single radiocardbon
determination and uncertainty and calculate the calendar age probability
density for it.
}
\examples{
CalibrateSingleDetermination(51020, 35, intcal20)
}
