% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/independent_calibration.R
\name{CalibrateSingleDetermination}
\alias{CalibrateSingleDetermination}
\title{Calibrate a single radiocarbon determination}
\usage{
CalibrateSingleDetermination(
  c14_determination,
  c14_uncertainty,
  calibration_data
)
}
\arguments{
\item{c14_determination}{A single observed radiocarbon determination (c14 age)}

\item{c14_uncertainty}{The uncertainty of the radiocarbon determination}

\item{calibration_data}{A dataframe which should contain one column entitled
c14_age and one column entitled c14_sig.
This format matches \link{intcal20}.}
}
\value{
A vector containing the probability for each calendar age in the
calibration data
}
\description{
Calibrate a single radiocarbon determination
}
\examples{
CalibrateSingleDetermination(51020, 35, intcal20)
}
