% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindPredictiveCalendarAgeDensity.R
\name{FindPredictiveCalendarAgeDensity}
\alias{FindPredictiveCalendarAgeDensity}
\title{Find the predictive calendar age density of from the output data}
\usage{
FindPredictiveCalendarAgeDensity(
  output_data,
  calendar_age_sequence,
  n_posterior_samples,
  interval_width = "2sigma",
  bespoke_probability = NA,
  n_burn = NA,
  n_end = NA
)
}
\arguments{
\item{output_data}{The return value from one of the updating functions e.g.
\link{WalkerBivarDirichlet} or
\link{PolyaUrnBivarDirichlet} or a list, each item containing
one of these values. Optionally, the output data can have an extra list item
named \code{label} which is used to set the label on the plot legend.}

\item{calendar_age_sequence}{A vector containing the calendar ages to
calculate the predictive density for.}

\item{n_posterior_samples}{Current number of samples it will draw from this
posterior to estimate the calendar age density (possibly repeats).}

\item{interval_width}{The confidence intervals to show for both the
calibration curve and the predictive density. Choose from one of \code{"1sigma"},
\code{"2sigma"} and \code{"bespoke"}. Default is \code{"2sigma"}.}

\item{bespoke_probability}{The probability to use for the confidence interval
if \code{"bespoke"} is chosen above. E.g. if 0.95 is chosen, then the 95\% confidence
interval is calculated. Ignored if \code{"bespoke"} is not chosen.}

\item{n_burn}{The number of samples required for burn-in - any samples before this
are not used to calculate the density. If not given, the first half of the
MCMC chain is discarded. Looking at the output of \link{PlotConvergenceData} can
help determine what an appropriate value is for a given output. Note that the maximum
value that can be chosen is \code{n_iter - 100 * n_thin} (where \code{n_iter} and \code{n_thin} are the
arguments given to \link{WalkerBivarDirichlet} or c\link{PolyaUrnBivarDirichlet}).}

\item{n_end}{The iteration number of the last sample to use. Assumed to be the number of iterations
if not given.}
}
\value{
A data frame of the \code{calendar_age}, the
\code{density_mean} and the confidence intervals for the density
\code{density_ci_lower} and \code{density_ci_upper}.
}
\description{
Given a calendar age sequence and output data from one of the package upate
functions, calculates the predicitive calendar age density and confidence
intervals. Note if you want to calculate and plot the result, use
\link{PlotPredictiveCalendarAgeDensity} instead.
}
\examples{
# Find results for example output, 2-sigma confidence interval (default)
FindPredictiveCalendarAgeDensity(
  walker_example_output, seq(600, 1700, length=12), 500)

# Find results for example output, 1-sigma confidence interval (default)
FindPredictiveCalendarAgeDensity(
  walker_example_output, seq(600, 1700, length=12), 500, "1sigma")

# Find results for example output, 95\% confidence interval (default)
FindPredictiveCalendarAgeDensity(
  walker_example_output, seq(600, 1700, length=12), 500, "bespoke", 0.95)
}
