% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindSummedProbabilityDistribution.R
\name{FindSummedProbabilityDistribution}
\alias{FindSummedProbabilityDistribution}
\title{Find the summed probability distribution (SPD) for a set of observations}
\usage{
FindSummedProbabilityDistribution(
  calendar_age_range,
  rc_determinations,
  rc_sigmas,
  F14C_inputs,
  calibration_curve
)
}
\arguments{
\item{calendar_age_range}{A vector of length 2 with the start and end
calendar age to calculate the SPD over}

\item{rc_determinations}{A vector of observed radiocarbon determinations}

\item{rc_sigmas}{A vector of the radiocarbon determinations
uncertainties (1-sigma). Must be the same length as \code{rc_determinations}.}

\item{F14C_inputs}{\code{TRUE} if the provided rc_determinations are F14C concentrations and \code{FALSE}
if they are radiocarbon age BP.}

\item{calibration_curve}{A dataframe which must contain one column \code{calendar_age_BP}, and also
columns \code{c14_age} and \code{c14_sig} or \code{f14c} and \code{f14c_sig}.
This format matches the curves supplied with this package e.g. \link{intcal20},
which contain all 5 columns.}
}
\value{
A data frame with one column \code{calendar_age_BP} containing the calendar
ages, and the other column \code{probability} containing the probability at that
calendar age
}
\description{
Takes a set of radiocarbon determinations and uncertainties and independently
calibrates each one, and then averages them to give the SPD estimate.
}
\examples{
FindSummedProbabilityDistribution(
   calendar_age_range=c(600, 1600),
   rc_determinations=c(602, 805, 1554),
   rc_sigmas=c(35, 34, 45),
   F14C_inputs=FALSE,
   calibration_curve=intcal20)
FindSummedProbabilityDistribution(
   calendar_age_range=c(600, 1600),
   rc_determinations=c(0.8, 0.85, 0.9),
   rc_sigmas=c(0.01, 0.015, 0.012),
   F14C_inputs=TRUE,
   calibration_curve=intcal20)
}
