% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{intcal04}
\alias{intcal04}
\title{IntCal04 calibration curve}
\format{
\subsection{\code{intcal04}}{

A data frame with 3,301 rows and 3 columns:
\describe{
\item{calendar_age}{The calendar age (yr BP)}
\item{c14_age}{The 14C age (yr BP)}
\item{c14_sig}{The uncertainty in the 14C age reported by the radiocarbon laboratory}
}
}
}
\source{
\url{https://doi.org/10.1017/S0033822200032999}
}
\usage{
intcal04
}
\description{
Atmospheric data from Reimer et al (2004)
Reimer et al. 2004
PJ Reimer, MGL Baillie, E Bard, A Bayliss, JW Beck, C Bertrand, PG Blackwell,
CE Buck, G Burr, KB Cutler, PE Damon, RL Edwards, RG Fairbanks, M Friedrich,
TP Guilderson, KA Hughen, B Kromer, FG McCormac, S Manning, C Bronk Ramsey,
RW Reimer, S Remmele, JR Southon, M Stuiver, S Talamo, FW Taylor,
J van der Plicht, and CE Weyhenmeyer (2004), Radiocarbon 46:1029-1058.
}
\keyword{datasets}
