% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{intcal13}
\alias{intcal13}
\title{IntCal13 calibration curve}
\format{
\subsection{\code{intcal13}}{

A data frame with 5,141 rows and 3 columns:
\describe{
\item{calendar_age}{The calendar age (yr BP)}
\item{c14_age}{The 14C age (yr BP)}
\item{c14_sig}{The uncertainty in the 14C age reported by the radiocarbon laboratory}
}
}
}
\source{
\url{http://doi.org/10.2458/azu_js_rc.55.16947}
}
\usage{
intcal13
}
\description{
Atmospheric data from Reimer et al (2013)
Reimer et al. 2013
Reimer PJ, Bard E, Bayliss A, Beck JW, Blackwell PG, Bronk Ramsey C, Buck CE,
Cheng H, Edwards RL, Friedrich M, Grootes PM, Guilderson TP, Haflidason H,
Hajdas I, Hatt? C, Heaton TJ, Hogg AG, Hughen KA, Kaiser KF, Kromer B,
Manning SW, Niu M, Reimer RW, Richards DA, Scott EM, Southon JR, Turney CSM,
van der Plicht J.
IntCal13 and MARINE13 radiocarbon age calibration curves 0-50000 years calBP
Radiocarbon 55(4). DOI: 10.2458/azu_js_rc.55.16947
}
\keyword{datasets}
