% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{intcal98}
\alias{intcal98}
\title{IntCal98 calibration curve}
\format{
\subsection{\code{intcal98}}{

A data frame with 1,538 rows and 5 columns:
\describe{
\item{calendar_age}{The calendar age (yr BP)}
\item{c14_age}{The 14C age (yr BP)}
\item{c14_sig}{The uncertainty in the 14C age reported by the radiocarbon laboratory}
\item{f14c}{The radiocarbon age expressed as F14C conctration}
\item{f14c_sig}{The uncertainty in F14C}
}
}
}
\source{
\url{https://doi.org/10.1017/S0033822200019123}
}
\usage{
intcal98
}
\description{
Atmospheric data from Stuiver et al 1998
Stuiver et al 1998
M. Stuiver, P. J. Reimer, E. Bard, J. W. Beck, G. S. Burr, K. A. Hughen,
B. Kromer, F. G. McCormac, J. v. d. Plicht and M. Spurk.
INTCAL98 Radiocarbon Age Calibration, 24,000-0 cal BP.
Radiocarbon 40, 1041--1083 (1998).
}
\keyword{datasets}
