% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterpolateCalibrationCurve.R
\name{InterpolateCalibrationCurve}
\alias{InterpolateCalibrationCurve}
\title{Interpolate calendar ages for a calibration curve}
\usage{
InterpolateCalibrationCurve(
  new_calendar_ages,
  calibration_curve,
  F14C_outputs = NA
)
}
\arguments{
\item{new_calendar_ages}{A scalar or vector containing calendar ages (in yr BP) to
interpolate the calibration curve to. If not provided (and \code{NA} is given) uses the range
from the minimum calendar age to the maximum calendar age of the original calibration curve
spaced by 1.}

\item{calibration_curve}{A dataframe which must contain one column \code{calendar_age_BP}, and also
columns \code{c14_age} and \code{c14_sig} or \code{f14c} and \code{f14c_sig}.
This format matches the curves supplied with this package e.g. \link{intcal20},
which contain all 5 columns.}

\item{F14C_outputs}{\code{TRUE} if only F14C concentrations are required, \code{FALSE}
if only radiocarbon age BP are required and \code{NA} if both are required for the new curve.}
}
\value{
A new dataframe with entries for the interpolated \code{c14_age}, and
\code{c14_sig}, \code{f14c} and \code{f14c_sig} values at the \code{calendar_age_BP} values given
in \code{new_calendar_ages}
}
\description{
Interpolate calendar ages for a calibration curve
}
\examples{
InterpolateCalibrationCurve(51020.5, intcal20)
InterpolateCalibrationCurve(c(51020.5, 51021.5), intcal20, TRUE)
InterpolateCalibrationCurve(c(51020.5, 51021.5), intcal20, FALSE)
InterpolateCalibrationCurve(NA, intcal20)
}
