% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_calendar_age_density.R
\name{PlotCalendarAgeDensity}
\alias{PlotCalendarAgeDensity}
\title{Plots the calendar age density of all objects from the output data}
\usage{
PlotCalendarAgeDensity(
  c14_determinations,
  c14_uncertainties,
  calibration_curve,
  output_data,
  n_posterior_samples,
  show_SPD = TRUE,
  show_confidence_intervals = TRUE,
  true_density = NULL,
  xlimscal = 1,
  ylimscal = 1,
  denscale = 3
)
}
\arguments{
\item{c14_determinations}{An array of observed radiocarbon determinations}

\item{c14_uncertainties}{An array of the radiocarbon determinations
uncertainties. Must be the same length as \code{c14_determinations}.}

\item{calibration_curve}{A dataframe which should contain at least 3 columns
entitled calendar_age, c14_age and c14_sig.
This format matches \link{intcal20}.}

\item{output_data}{The return value from one of the updating functions e.g.
\link{WalkerBivarDirichlet} or
\link{BivarGibbsDirichletwithSlice} or a list, each item containing
one of these values. Optionally, the output data can have an extra list item
named \code{label} which is used to set the label on the plot legend.}

\item{n_posterior_samples}{Current number of samples it will draw from this
posterior to estimate the calendar age density (possibly repeats).}

\item{show_SPD}{Whether to calculate and show the summed probability function
on the plot (optional). Default is \code{TRUE}.}

\item{show_confidence_intervals}{Whether to show the 95\% confidence intervals
of the posterior density on the plot. Default is \code{TRUE}.}

\item{true_density}{The true calendar age density (optional). Expects a data
frame with two columns, the first column being calendar age and the right
column being the normalized density.}

\item{xlimscal, ylimscal}{Whether to scale the x or y limits (optional).
Default is 1. Values more than 1 will increase range of the limits,
values less than 1 will decrease the range of the limits.}

\item{denscale}{Whether to scale the vertical range of the density plot
relative to the calibration curve plot (optional). Default is 3 which means
that the maximum SPD density will be at 1/3 of the height of the plot.
TODO: base on the output data.}
}
\value{
A list, each item containing a data frame of the \code{calendar_age}, the
\code{density_mean} and the 95\% confidence intervals for the density
\code{density_ci_lower} and \code{density_ci_upper} for each set of output data.
}
\description{
Plots the input radiocarbon determinations and calibration curve, with the
output predicted density on the same plot. Can also optionally show the
SPD estimate
}
\examples{
# Plot results for a single calibration
PlotCalendarAgeDensity(
  c14_determinations = kerr$c14_ages,
  c14_uncertainties = kerr$c14_sig,
  calibration_curve = intcal20,
  output_data = walker_example_output,
  n_posterior_samples = 500)

# Plot results from a calibration, and add a label
new_output = walker_example_output
new_output$label = "My plot"
PlotCalendarAgeDensity(
  c14_determinations = kerr$c14_ages,
  c14_uncertainties = kerr$c14_sig,
  calibration_curve = intcal20,
  output_data = walker_example_output,
  n_posterior_samples = 500)
}
