% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotCalendarAgeDensityIndividualSample.R
\name{PlotCalendarAgeDensityIndividualSample}
\alias{PlotCalendarAgeDensityIndividualSample}
\title{Plots the posterior calendar ages for an individual determination}
\usage{
PlotCalendarAgeDensityIndividualSample(
  ident,
  output_data,
  calibration_curve = NULL,
  n_breaks = NA,
  interval_width = "2sigma",
  bespoke_probability = NA
)
}
\arguments{
\item{ident}{the determination you want to show the individual posterior
calendar age for.}

\item{output_data}{The return value from one of the updating functions e.g.
\link{WalkerBivarDirichlet} or
\link{PolyaUrnBivarDirichlet} or a list, each item containing
one of these values. Optionally, the output data can have an extra list item
named \code{label} which is used to set the label on the plot legend.}

\item{calibration_curve}{This is usually not required since the name of the
calibration curve variable is saved in the output data. However if the
variable with this name is no longer in your environment then you should pass
the calibration curve here. If provided this should be a dataframe which
should contain at least 3 columns entitled calendar_age, c14_age and c14_sig.
This format matches \link{intcal20}.}

\item{n_breaks}{The number of breaks in the histogram (optional). If not
given then it is set to the max of 100 and 1/tenth of post burn-in chain.}

\item{interval_width}{The confidence intervals to show for the
calibration curve. Choose from one of "1sigma", "2sigma" and "bespoke".
Default is "2sigma".}

\item{bespoke_probability}{The probability to use for the confidence interval
if "bespoke" is chosen above. E.g. if 0.95 is chosen, then the 95\% confidence
interval is calculated. Ignored if "bespoke" is not chosen.}
}
\value{
No return value
}
\description{
Once a function has been run to calibrate a set of radiocarbon
determinations, the posterior density for a single determination can be
plotted using this function.
}
\examples{
# Plot results for the 10th determinations
PlotCalendarAgeDensityIndividualSample(10, walker_example_output)
}
