% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotCalendarAgeDensityIndividualSample.R
\name{PlotCalendarAgeDensityIndividualSample}
\alias{PlotCalendarAgeDensityIndividualSample}
\title{Plots the posterior calendar ages for an individual determination}
\usage{
PlotCalendarAgeDensityIndividualSample(
  ident,
  c14_determinations,
  c14_sigmas,
  calibration_curve,
  output_data,
  n_breaks = NA
)
}
\arguments{
\item{ident}{the determination you want to show the individual posterior
calendar age for.}

\item{c14_determinations}{An array of observed radiocarbon determinations}

\item{c14_sigmas}{An array of the radiocarbon determinations
uncertainties. Must be the same length as \code{c14_determinations}.}

\item{calibration_curve}{A dataframe which should contain at least 3 columns
entitled calendar_age, c14_age and c14_sig.
This format matches \link{intcal20}.}

\item{output_data}{The return value from one of the updating functions e.g.
\link{WalkerBivarDirichlet} or
\link{PolyaUrnBivarDirichlet} or a list, each item containing
one of these values. Optionally, the output data can have an extra list item
named \code{label} which is used to set the label on the plot legend.}

\item{n_breaks}{The number of breaks in the histogram (optional). If not
given then it is set to the max of 100 and 1/tenth of post burn-in chain.}
}
\value{
No return value
}
\description{
Once a function has been run to calibrate a set of radiocarbon
determinations, the posterior density for a single determination can be
plotted using this function.
}
\examples{
# Plot results for the 10th determinations
PlotCalendarAgeDensityIndividualSample(
  ident = 10,
  c14_determinations = kerr$c14_ages,
  c14_sigmas = kerr$c14_sig,
  calibration_curve = intcal20,
  output_data = walker_example_output)

}
