% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotCalendarAgeDensityIndividualSample.R
\name{PlotCalendarAgeDensityIndividualSample}
\alias{PlotCalendarAgeDensityIndividualSample}
\title{Plots the posterior calendar ages for an individual determination}
\usage{
PlotCalendarAgeDensityIndividualSample(
  ident,
  output_data,
  calibration_curve = NULL,
  plot_14C_age = TRUE,
  resolution = 5,
  interval_width = "2sigma",
  bespoke_probability = NA,
  n_burn = NA
)
}
\arguments{
\item{ident}{the determination you want to show the individual posterior
calendar age for.}

\item{output_data}{The return value from one of the updating functions e.g.
\link{WalkerBivarDirichlet} or
\link{PolyaUrnBivarDirichlet} or a list, each item containing
one of these values. Optionally, the output data can have an extra list item
named \code{label} which is used to set the label on the plot legend.}

\item{calibration_curve}{This is usually not required since the name of the
calibration curve variable is saved in the output data. However if the
variable with this name is no longer in your environment then you should pass
the calibration curve here. If provided this should be a dataframe which
should contain at least 3 columns entitled calendar_age, c14_age and c14_sig.
This format matches \link{intcal20}.}

\item{plot_14C_age}{Whether to use the 14C yr BP as the units of the y-axis.
Defaults to TRUE. If FALSE uses F14C concentration instead.}

\item{resolution}{The distance between histogram breaks for the calendar age density.
Must be an integer greater than one.}

\item{interval_width}{The confidence intervals to show for the
calibration curve. Choose from one of "1sigma", "2sigma" and "bespoke".
Default is "2sigma".}

\item{bespoke_probability}{The probability to use for the confidence interval
if "bespoke" is chosen above. E.g. if 0.95 is chosen, then the 95\% confidence
interval is calculated. Ignored if "bespoke" is not chosen.}

\item{n_burn}{The number of samples required for burn-in - any samples before this
are not used to calculate the density. If not given, the first half of the
MCMC chain is discarded. Looking at the output of \link{PlotConvergenceData} can
help determine what an appropriate value is for a given output. Note that the maximum
value that can be chosen is \code{n_iter - 100 * n_thin} (where \code{n_iter} and \code{n_thin} are the
arguments given to \link{WalkerBivarDirichlet} or c\link{PolyaUrnBivarDirichlet}).}
}
\value{
No return value
}
\description{
Once a function has been run to calibrate a set of radiocarbon
determinations, the posterior density for a single determination can be
plotted using this function.
}
\examples{
# Plot results for the 10th determinations
PlotCalendarAgeDensityIndividualSample(10, walker_example_output)
}
