% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotConvergenceData.R
\name{PlotConvergenceData}
\alias{PlotConvergenceData}
\title{Plots data that can be used to estimate convergence of the MCMC sampler}
\usage{
PlotConvergenceData(output_data, n_initial = 1000)
}
\arguments{
\item{output_data}{The return value from one of the updating functions e.g.
\link{WalkerBivarDirichlet} or
\link{PolyaUrnBivarDirichlet}.}

\item{n_initial}{The number of samples to use for the 'initial' predictive density,
which is then compared with all subsequent data points. If not given 1000 data points will
be used.}
}
\value{
No value
}
\description{
This plots the Kullback–Leibler divergence between the initial predictive density
and the density calculated during a run of a calibration function. This can be used to
estimate whether the calibration function has converged.
}
\examples{
# Plot results for the example data
PlotConvergenceData(walker_example_output, 500)
}
