% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Convergence.R
\name{PlotGelmanRubinDiagnosticMultiChain}
\alias{PlotGelmanRubinDiagnosticMultiChain}
\title{Plots a histogram of the Gelman–Rubin convergence diagnostic for a single chain}
\usage{
PlotGelmanRubinDiagnosticMultiChain(
  output_data_1,
  output_data_2,
  output_data_3,
  n_burn
)
}
\arguments{
\item{output_data_1}{The return value from one of the updating functions e.g.
\link{WalkerBivarDirichlet} or
\link{PolyaUrnBivarDirichlet}.}

\item{n_burn}{The number of samples required for burn-in.}
}
\value{
No value
}
\description{
This plots a histogram of the potential scale reduction factor (PSRF) for each of
the calendar age observations for a single MCMC chain, by splitting the chain into 3 parts after
n_burn and comparing the within-chain variance with the between-chains variance.
If the chain have converged to the target posterior distribution, then PSRF should be close to 1.
}
\examples{
# Plot results for the example data
PlotConvergenceData(walker_example_output, 500)
}
