% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_individual_calendar_age_density.R
\name{PlotIndividualCalendarAgeDensity}
\alias{PlotIndividualCalendarAgeDensity}
\title{Plots the posterior calendar ages for an individual determination}
\usage{
PlotIndividualCalendarAgeDensity(
  ident,
  c14_determinations,
  c14_uncertainties,
  calibration_curve,
  output_data,
  n_breaks = NA
)
}
\arguments{
\item{ident}{the determination you want to show the individual posterior
calendar age for.}

\item{c14_determinations}{A vector containing the radiocarbon determinations.}

\item{c14_uncertainties}{A vector containing the radiocarbon determination
uncertainties. Must be the same length as \code{c14_determinations}.}

\item{calibration_curve}{A dataframe which should contain one column entitled
c14_age and one column entitled c14_sig.
This format matches \link{intcal20}.}

\item{output_data}{Data returned from one of the updating functions e.g.
\link{WalkerBivarDirichlet} or
\link{BivarGibbsDirichletwithSlice}.}

\item{n_breaks}{The number of breaks in the histogram (optional). If not
given then it is set to the max of 100 and 1/tenth of post burn-in chain.}
}
\value{
No return value
}
\description{
Once a function has been run to calibrate a set of radiocarbon
determinations, the posterior density for a single determination can be
plotted using this function.
}
\examples{
# First generate some output data
c14_determinations = c(602, 805, 954)
c14_uncertainties = c(35, 34, 45)
walker_temp = WalkerBivarDirichlet(
  c14_determinations = c14_determinations,
  c14_uncertainties = c14_uncertainties,
  calibration_curve = intcal20,
  lambda = 0.1,
  nu1 = 0.25,
  nu2 = 10,
  alpha_shape = 1,
  alpha_rate = 1)

# Plot results for the second determinations
PlotIndividualCalendarAgeDensity(
  ident = 2,
  c14_determinations = c14_determinations,
  c14_uncertainties = c14_uncertainties,
  calibration_curve = intcal20,
  output_data = walker_temp,
  n_breaks = 10)

}
