% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_number_of_clusters.R
\name{PlotNumberOfClusters}
\alias{PlotNumberOfClusters}
\title{Plots the number of clusters}
\usage{
PlotNumberOfClusters(output_data)
}
\arguments{
\item{output_data}{Data returned from one of the updating functions e.g.
\link{WalkerBivarDirichlet} or
\link{BivarGibbsDirichletwithSlice}.}
}
\value{
No return value
}
\description{
Plots the number of clusters
}
