% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_number_of_clusters.R
\name{PlotNumberOfClusters}
\alias{PlotNumberOfClusters}
\title{Plots the number of clusters}
\usage{
PlotNumberOfClusters(output_data)
}
\arguments{
\item{output_data}{Data returned from one of the updating functions e.g.
\link{WalkerBivarDirichlet} or
\link{BivarGibbsDirichletwithSlice}.}
}
\value{
No return value
}
\description{
Once a function has been run to calibrate a set of radiocarbon
determinations, the estimated number of clusters can be plotted using this
function.
}
\examples{
# First generate some output data
walker_temp = WalkerBivarDirichlet(
  c14_determinations = c(602, 805, 954),
  c14_uncertainties = c(35, 34, 45),
  calibration_curve = intcal20,
  lambda = 0.1,
  nu1 = 0.25,
  nu2 = 10,
  alpha_shape = 1,
  alpha_rate = 1)

# Plot results for the number of clusters
PlotNumberOfClusters(output_data = walker_temp)
}
