% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_number_of_clusters.R
\name{PlotNumberOfClusters}
\alias{PlotNumberOfClusters}
\title{Plots the number of clusters}
\usage{
PlotNumberOfClusters(output_data)
}
\arguments{
\item{output_data}{The return value from one of the updating functions e.g.
\link{WalkerBivarDirichlet} or
\link{BivarGibbsDirichletwithSlice} or a list, each item containing
one of these values. Optionally, the output data can have an extra list item
named \code{label} which is used to set the label on the plot legend.}
}
\value{
No return value
}
\description{
Once a function has been run to calibrate a set of radiocarbon
determinations, the estimated number of clusters can be plotted using this
function.
}
\examples{
# Plot results for the number of clusters
PlotNumberOfClusters(output_data = walker_example_output)
}
