% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotNumberOfClusters.R
\name{PlotNumberOfClusters}
\alias{PlotNumberOfClusters}
\title{Plots the number of clusters}
\usage{
PlotNumberOfClusters(output_data, n_burn = NA, n_end = NA)
}
\arguments{
\item{output_data}{The return value from one of the updating functions e.g.
\link{PolyaUrnBivarDirichlet} or
\link{WalkerBivarDirichlet} or a list, each item containing
one of these values. Optionally, the output data can have an extra list item
named \code{label} which is used to set the label on the plot legend.}

\item{n_burn}{The number of samples required for burn-in - any samples before this
are not used in the calculation. If not given, the first half of the
MCMC chain is discarded. Looking at the output of \link{PlotConvergenceData} can
help determine what an appropriate value is for a given output. Note that the maximum
value that can be chosen is \code{n_iter - 100 * n_thin} (where \code{n_iter} and \code{n_thin} are the
arguments given to \link{PolyaUrnBivarDirichlet} or \link{WalkerBivarDirichlet}).}

\item{n_end}{The iteration number of the last sample to use. Assumed to be the number of iterations
if not given.}
}
\value{
None
}
\description{
Once a function has been run to calibrate a set of radiocarbon
determinations, the estimated number of clusters can be plotted using this
function.
}
\examples{
# Plot results for the number of clusters
polya_urn_output <- PolyaUrnBivarDirichlet(
    two_normals$c14_age, two_normals$c14_sig, intcal20, n_iter = 1e4, show_progress = FALSE)
PlotNumberOfClusters(polya_urn_output)
}
