% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotNumberOfInternalChanges.R
\name{PlotNumberOfInternalChanges}
\alias{PlotNumberOfInternalChanges}
\title{Plots the number of internal changes}
\usage{
PlotNumberOfInternalChanges(output_data, n_burn = NA, n_end = NA)
}
\arguments{
\item{output_data}{The return value from the updating functions
\link{PPcalibrate}. Optionally, the output data can have an extra list item
named \code{label} which is used to set the label on the plot legend.}

\item{n_burn}{The number of samples required for burn-in - any samples before this
are not used in the calculation. If not given, the first half of the
MCMC chain is discarded. Note that the maximum
value that can be chosen is \code{n_iter - 100 * n_thin} (where \code{n_iter} and \code{n_thin} are the
arguments given to \link{PPcalibrate}).}

\item{n_end}{The iteration number of the last sample to use. Assumed to be the number of iterations
if not given.}
}
\value{
No return value
}
\description{
Once a function has been run to calibrate a set of radiocarbon determinations,
the estimated number of internal changes can be plotted using this function.
}
\examples{
# TODO
}
