% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPosteriorHeights.R
\name{PlotPosteriorHeights}
\alias{PlotPosteriorHeights}
\title{Plots the posterior heights}
\usage{
PlotPosteriorHeights(
  output_data,
  n_changes = c(1, 2, 3),
  n_burn = NA,
  n_end = NA,
  kernel_bandwidth = NA
)
}
\arguments{
\item{output_data}{The return value from the updating functions
\link{PPcalibrate}. Optionally, the output data can have an extra list item
named \code{label} which is used to set the label on the plot legend.}

\item{n_changes}{Which number of internal changes to plot for - a vector which can contain at most 4 elements, with
values in the range 1 to 6. If not given \code{c(1, 2, 3)} will be used.}

\item{n_burn}{The number of samples required for burn-in - any samples before this
are not used in the calculation. If not given, the first half of the
MCMC chain is discarded. Note that the maximum
value that can be chosen is \code{n_iter - 100 * n_thin} (where \code{n_iter} and \code{n_thin} are the
arguments given to \link{PPcalibrate}).}

\item{n_end}{The iteration number of the last sample to use. Assumed to be the number of iterations
if not given.}

\item{kernel_bandwidth}{The bandwidth used for the KDE of the density (optional). If not give 1/50th of the
maximum height will be used.}
}
\description{
Description here
}
\examples{
# Plot results for a single calibration
# TODO
}
