% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPredictiveCalendarAgeDensity.R
\name{PlotPredictiveCalendarAgeDensity}
\alias{PlotPredictiveCalendarAgeDensity}
\title{Plots the calendar age density of all objects from the output data}
\usage{
PlotPredictiveCalendarAgeDensity(
  output_data,
  n_posterior_samples,
  calibration_curve = NULL,
  plot_14C_age = TRUE,
  show_SPD = FALSE,
  show_confidence_intervals = TRUE,
  interval_width = "2sigma",
  bespoke_probability = NA,
  denscale = 3,
  n_calc = 1001,
  n_burn = NA,
  n_end = NA
)
}
\arguments{
\item{output_data}{The return value from one of the updating functions e.g.
\link{WalkerBivarDirichlet} or
\link{PolyaUrnBivarDirichlet} or a list, each item containing
one of these values. Optionally, the output data can have an extra list item
named \code{label} which is used to set the label on the plot legend.}

\item{n_posterior_samples}{Current number of samples it will draw from this
posterior to estimate the calendar age density (possibly repeats).}

\item{calibration_curve}{This is usually not required since the name of the
calibration curve variable is saved in the output data. However if the
variable with this name is no longer in your environment then you should pass
the calibration curve here. If provided this should be a dataframe which
should contain at least 3 columns entitled calendar_age, c14_age and c14_sig.
This format matches \link{intcal20}.}

\item{plot_14C_age}{Whether to use the 14C yr BP as the units of the y-axis.
Defaults to TRUE. If FALSE uses F14C concentration instead.}

\item{show_SPD}{Whether to calculate and show the summed probability
distribution on the plot (optional). Default is \code{FALSE}.}

\item{show_confidence_intervals}{Whether to show the confidence intervals
for the chosen probability on the plot. Default is \code{TRUE}.}

\item{interval_width}{The confidence intervals to show for both the
calibration curve and the predictive density. Choose from one of \code{"1sigma"},
\code{"2sigma"} and \code{"bespoke"}. Default is \code{"2sigma"}.}

\item{bespoke_probability}{The probability to use for the confidence interval
if \code{"bespoke"} is chosen above. E.g. if 0.95 is chosen, then the 95\% confidence
interval is calculated. Ignored if \code{"bespoke"} is not chosen.}

\item{denscale}{Whether to scale the vertical range of the density plot
relative to the calibration curve plot (optional). Default is 3 which means
that the maximum SPD density will be at 1/3 of the height of the plot.}

\item{n_calc}{Number of points to use when calculating the predictive
density. Default is 1001.}

\item{n_burn}{The number of samples required for burn-in - any samples before this
are not used in the calculation. If not given, the first half of the
MCMC chain is discarded. Looking at the output of \link{PlotConvergenceData} can
help determine what an appropriate value is for a given output. Note that the maximum
value that can be chosen is \code{n_iter - 100 * n_thin} (where \code{n_iter} and \code{n_thin} are the
arguments given to \link{PolyaUrnBivarDirichlet} or c\link{WalkerBivarDirichlet}).}

\item{n_end}{The iteration number of the last sample to use. Assumed to be the number of iterations
if not given.}
}
\value{
A list, each item containing a data frame of the \code{calendar_age}, the
\code{density_mean} and the 95\% confidence intervals for the density
\code{density_ci_lower} and \code{density_ci_upper} for each set of output data.
}
\description{
Plots the input radiocarbon determinations and calibration curve, with the
output predicted density on the same plot. Can also optionally show the
SPD estimate. Note that if all you are only interested in is the density
data, without an accompanying plot, you can use
\link{FindPredictiveCalendarAgeDensity} instead.
}
\examples{
# Plot results for a single calibration
PlotPredictiveCalendarAgeDensity(polya_urn_example_output, 500)

# Plot results from a calibration, and add a label
op = PolyaUrnBivarDirichlet(two_normals$c14_age, two_normals$c14_sig, intcal20, n_iter=1e4)
op$label = "My plot"
PlotPredictiveCalendarAgeDensity(op, 500)

# Plot results from two calibrations on the same plot, and show the SPD
PlotPredictiveCalendarAgeDensity(
  list(walker_example_output, polya_urn_example_output), 500, show_SPD = TRUE)

# Plot and show the 1-sigma confidence interval
PlotPredictiveCalendarAgeDensity(polya_urn_example_output, 500, interval_width = "1sigma")

# Plot and show the 80\% confidence interval
PlotPredictiveCalendarAgeDensity(
  polya_urn_example_output, 500, interval_width = "bespoke", bespoke_probability = 0.8)
}
