% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolyaUrnBivarDirichlet.R
\name{PolyaUrnBivarDirichlet}
\alias{PolyaUrnBivarDirichlet}
\title{Calibration of a set of individual radiocarbon samples by performing the
Gibbs sampler using a DPMM}
\usage{
PolyaUrnBivarDirichlet(
  rc_determinations,
  rc_sigmas,
  calibration_curve,
  F14C_inputs = FALSE,
  n_iter = 1e+05,
  n_thin = 10,
  use_F14C_space = TRUE,
  slice_width = NA,
  slice_multiplier = 10,
  n_clust = min(10, length(rc_determinations)),
  show_progress = TRUE,
  sensible_initialisation = TRUE,
  lambda = NA,
  nu1 = NA,
  nu2 = NA,
  A = NA,
  B = NA,
  alpha_shape = NA,
  alpha_rate = NA,
  mu_phi = NA,
  calendar_ages = NA
)
}
\arguments{
\item{rc_determinations}{A vector of observed radiocarbon determinations}

\item{rc_sigmas}{A vector of the radiocarbon determinations
uncertainties (1-sigma). Must be the same length as \code{rc_determinations}.}

\item{calibration_curve}{A dataframe which must contain one column \code{calendar_age_BP}, and also
columns \code{c14_age} and \code{c14_sig} or \code{f14c} and \code{f14c_sig} (or both sets).
This format matches the curves supplied with this package e.g. \link{intcal20}, \link{intcal13},
which contain all 5 columns.}

\item{F14C_inputs}{\code{TRUE} if the provided rc_determinations are F14C concentrations and \code{FALSE}
if they are radiocarbon age BP. Defaults to \code{FALSE}.}

\item{n_iter}{The number of MCMC iterations (optional). Default is 100,000.}

\item{n_thin}{How much to thin the output (optional). 1 is no thinning,
a larger number is more thinning. Default is 10. Must choose an integer more
than 1 and not too close to \code{n_iter}, to ensure there are enought samples from
posterior to potentially use.}

\item{use_F14C_space}{Whether the calculations are carried out in F14C space (default is TRUE).
If FALSE, calculations are carried out in \eqn{{}^{14}}C yr BP space.}

\item{slice_width}{Parameter for slice sampling (optional). If not given a value
is chosen intelligently based on the spread of the initial calendar ages.
Must be given if \code{sensible_initialisation} is \code{FALSE}.}

\item{slice_multiplier}{Integer parameter for slice sampling (optional).
Default is 10. Limits the slice size to \code{slice_multiplier * slice_width}.}

\item{n_clust}{The initial number of clusters (optional). Must
be less than the length of \code{rc_determinations}. Default is 10 or the length
of \code{rc_determinations} if it is less than 10.}

\item{show_progress}{Whether to show a progress bar in the console during
execution. Default is \code{TRUE}.}

\item{sensible_initialisation}{Whether to use sensible start values and
adaptive prior on \eqn{\mu_{\phi}} and  (A, B).
If this is \code{TRUE} (the default), then all the remaining arguments below are
ignored.}

\item{lambda, nu1, nu2}{Hyperparameters for the prior on the means
\eqn{\phi_j} and precision \eqn{\tau_j} of each individual calendar age
cluster \eqn{j}.
\deqn{(\phi_j, \tau_j)|\mu_{\phi} \sim
\textrm{NormalGamma}(\mu_{\phi}, \lambda, \nu_1, \nu_2)} where
\eqn{\mu_{\phi}} is the overall cluster centering. Required if
\code{sensible_initialisation} is \code{FALSE}.}

\item{A, B}{Prior on \eqn{\mu_{\phi}} giving the mean and precision of the
overall centering \eqn{\mu_{\phi} \sim N(A, B^{-1})} i.e.
B small is uninformative. Required if \code{sensible_initialisation} is \code{FALSE}.}

\item{alpha_shape, alpha_rate}{Hyperparameters for the shape and rate on prior
for DP concentration, \eqn{\alpha}, determining the number of clusters we
expect to observe among our n sampled objects.
\eqn{\alpha \sim \Gamma(\eta_1, \eta_2)} where \eqn{\eta_1, \eta_2} are
the \code{alpha_shape} and \code{alpha_rate}. A small alpha means more concentrated
(i.e. few clusters) while a large alpha means not concentrated (i.e. many
clusters).  Required if \code{sensible_initialisation} is \code{FALSE}.}

\item{mu_phi}{Initial value of the overall cluster centering \eqn{\mu_{\phi}}.
Required if \code{sensible_initialisation} is \code{FALSE}.}

\item{calendar_ages}{The initial estimate for the underlying calendar ages
(optional). If supplied it must be a vector with the same length as
\code{rc_determinations}.  Required if \code{sensible_initialisation} is \code{FALSE}.}
}
\value{
A list with 10 items. The first 7 items contain output data, each of
which have one dimension of size \eqn{n_{\textrm{out}} =
\textrm{floor}( n_{\textrm{iter}}/n_{\textrm{thin}}) + 1}, each row storing
the result from every \eqn{n_{\textrm{thin}}}th iteration:

\describe{
\item{\code{cluster_identifiers}}{A list of length \eqn{n_{\textrm{out}}} each entry
giving the cluster allocation - an integer between 1 and n_clust - for each
observation.}
\item{\code{alpha}}{A double vector of length \eqn{n_{\textrm{out}}} giving the
DP concentration parameter.}
\item{\code{n_clust}}{An integer vector of length \eqn{n_{\textrm{out}}} giving
the number of clusters.}
\item{\code{phi}}{A list of length \eqn{n_{\textrm{out}}} each entry giving
a vector of length n_clust of the cluster means \eqn{\phi_j}.}
\item{\code{tau}}{A list of length \eqn{n_{\textrm{out}}} each entry giving
a vector of length n_clust of the cluster uncertainties \eqn{\tau_j}.}
\item{\code{observations_per_cluster}}{A list of length \eqn{n_{\textrm{out}}} each entry giving
a vector of length n_clust of the number of observations for that cluster.}
\item{\code{calendar_ages}}{An \eqn{n_{\textrm{out}}} by \eqn{n_{\textrm{obs}}}
integer matrix. Gives the calendar age for each observation.}
\item{\code{mu_phi}}{A vector of length \eqn{n_{\textrm{out}}} giving the overall
centering \eqn{\mu_{\phi}} of the clusters.}
}
where \eqn{n_{\textrm{obs}}} is the number of radiocarbon observations i.e.
the length of \code{rc_determinations}.

The remaining items give information about input data, input parameters (or
those calculated using \code{sensible_initialisation}) and update_type

\describe{
\item{\code{update_type}}{A string that always has the value "Polya Urn".}
\item{\code{input_data}}{a list containing the C14 data used and the name of
the calibration curve used.}
\item{\code{input_parameters}}{A list containing the values of the fixed
hyperparameters \code{lambda}, \code{nu1}, \code{nu2}, \code{A}, \code{B}, \code{alpha_shape},
\code{alpha_rate} and \code{mu_phi}, and the slice parameters \code{slice_width} and
\code{slice_multiplier}.}
}
}
\description{
This function takes as an input a set of radiocarbon
determinations and associated 1-sigma uncertainties, as well as the
calibration curve which should be used, and returns output data that can be
sampled to estimate the joint calendar age density and clusters. This method
considers both the mean and the variance of the clusters to be unknown.

For more information read the vignette: \cr
\code{vignette("Non-parametric-summed-density", package = "carbondate")}
}
\examples{
# Note these examples are shown with a small n_iter to speed up execution.
# When you run ensure n_iter gives convergence (try function default).

# Basic usage making use of sensible initialisation to set most values and
# using a saved example data set and the IntCal20 curve.
polya_urn_output <- PolyaUrnBivarDirichlet(
    two_normals$c14_age,
    two_normals$c14_sig,
    intcal20,
    n_iter = 100,
    show_progress = FALSE)

# The radiocarbon determinations can be given as F14C concentrations
polya_urn_output <- PolyaUrnBivarDirichlet(
    two_normals$f14c,
    two_normals$f14c_sig,
    intcal20,
    F14C_inputs = TRUE,
    n_iter = 100,
    show_progress = FALSE)
}
