% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pp_uniform_phase}
\alias{pp_uniform_phase}
\title{Example artificial data - Uniform Phase}
\format{
\subsection{\code{pp_uniform_phase}}{

A data frame with 40 rows and 4 columns:
\describe{
\item{c14_age}{The 14C age (yr BP)}
\item{c14_sig}{The uncertainty in the 14C age}
\item{f14c}{The F14C concentration}
\item{f14c_sig}{The uncertainty in the F14C concentration}
}
}
}
\usage{
pp_uniform_phase
}
\description{
40 radiocarbon determinations for which the underlying calendar ages are
drawn .... Tim to complete
}
\keyword{datasets}
