% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pp_uniform_phase}
\alias{pp_uniform_phase}
\title{Example artificial data - Uniform Phase}
\format{
\subsection{\code{pp_uniform_phase}}{

A data frame with 40 rows and 4 columns:
\describe{
\item{c14_age}{The 14C age (yr BP)}
\item{c14_sig}{The uncertainty in the 14C age}
\item{f14c}{The F14C concentration}
\item{f14c_sig}{The uncertainty in the F14C concentration}
}
}
}
\usage{
pp_uniform_phase
}
\description{
40 simulated radiocarbon determinations for which the underlying calendar ages are
drawn (uniformly at random) from the period  550--500 cal yr BP. The observational
uncertainty of each determination is set to be 15 \eqn{{}^{14}}C yrs. \cr \cr The corresponding
\eqn{{}^{14}}C ages are then simulated based upon the IntCal20 calibration curve (convolved with
the 15 \eqn{{}^{14}}C yr measurement uncertainty). This dataset matches that used in the
package vignette to illustrate the Poisson process modelling.
}
\keyword{datasets}
