% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shcal13}
\alias{shcal13}
\title{SHCal13 calibration curve}
\format{
\subsection{\code{shcal13}}{

A data frame with 5,141 rows and 5 columns:
\describe{
\item{calendar_age}{The calendar age (yr BP)}
\item{c14_age}{The \eqn{{}^{14}}C age (yr BP)}
\item{c14_sig}{The uncertainty in the \eqn{{}^{14}}C age reported by the radiocarbon laboratory}
\item{f14c}{The radiocarbon age expressed as F14C conctration}
\item{f14c_sig}{The uncertainty in F14C}
}
}
}
\source{
\url{http://doi.org/10.2458/azu_js_rc.55.16783}
}
\usage{
shcal13
}
\description{
SHCal13 atmospheric curve (Hogg et al 2013)
Hogg et al. 2013
Alan G Hogg, Quan Hua, Paul G Blackwell, Caitlin E Buck, Thomas P Guilderson,
Timothy J  Heaton, Mu Niu, Jonathan G Palmer, Paula J Reimer, Ron W Reimer,
Christian S M Turney, Susan R H Zimmerman.
2013
SHCal13 Southern Hemisphere Calibration, 0-50,000 Years cal BP.
Radiocarbon 55(4):1889-1903. doi:10.2458/azu_js_rc.55.16783
}
\keyword{datasets}
