% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{two_normals}
\alias{two_normals}
\title{Example artificial data - Mixture of Normal Phases}
\format{
\subsection{\code{two_normals}}{

A data frame with 50 rows and 4 columns:
\describe{
\item{c14_age}{The \eqn{{}^{14}}C age (yr BP)}
\item{c14_sig}{The uncertainty in the \eqn{{}^{14}}C age}
\item{f14c}{The F14C concentration}
\item{f14c_sig}{The uncertainty in the F14C concentration}
}
}
}
\usage{
two_normals
}
\description{
50 radiocarbon determinations for which the underlying calendar ages are
drawn from a mixture of two normals. This dataset is included simply to give
some quick-to-run examples.
}
\keyword{datasets}
