% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WalkerBivarDirichlet.R
\name{WalkerBivarDirichlet}
\alias{WalkerBivarDirichlet}
\title{Implements calibration using Walker updating of the DPMM}
\usage{
WalkerBivarDirichlet(
  x,
  xsig,
  calibration_curve,
  lambda,
  nu1,
  nu2,
  A,
  B,
  cprshape,
  cprrate,
  niter = 100,
  nthin = 10,
  theta = NA,
  slicew = 1000,
  m = 10,
  kstar = 10,
  sensibleinit = TRUE,
  showprogress = TRUE
)
}
\arguments{
\item{x}{A vector containing the radiocarbon determinations}

\item{xsig}{A vector containing the radiocarbon determination uncertainties.
Must be the same length as \code{x}.}

\item{calibration_curve}{A dataframe which should contain one column entitled
c14_age and one column entitled c14_sig.
This format matches \link{intcal20}.}

\item{lambda}{TODO}

\item{nu1}{TODO}

\item{nu2}{TODO}

\item{A}{TODO}

\item{B}{TODO}

\item{cprshape}{TODO}

\item{cprrate}{TODO}

\item{niter}{TODO}

\item{nthin}{TODO}

\item{theta}{TODO}

\item{slicew}{TODO}

\item{m}{TODO}

\item{kstar}{TODO}

\item{sensibleinit}{TODO}

\item{showprogress}{TODO}
}
\value{
a list TODO add more
}
\description{
Implements calibration using Walker updating of the DPMM
}
\examples{
WalkerBivarDirichlet(
  x=c(602, 805, 1554),
  xsig=c(35, 34, 45),
  calibration_curve=intcal20,
  lambda=1,
  nu1=1,
  nu2=1,
  A=1,
  B=1,
  cprshape=1,
  cprrate=1)
}
