% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WalkerBivarDirichlet.R
\name{WalkerBivarDirichlet}
\alias{WalkerBivarDirichlet}
\title{Calibration of a set of individual radiocarbon samples using Walker updating
of the DPMM}
\usage{
WalkerBivarDirichlet(
  c14_determinations,
  c14_sigmas,
  calibration_curve,
  n_iter = 100,
  n_thin = 10,
  slice_width = max(1000, diff(range(c14_determinations))/2),
  slice_multiplier = 10,
  show_progress = TRUE,
  sensible_initialisation = TRUE,
  lambda = NA,
  nu1 = NA,
  nu2 = NA,
  A = NA,
  B = NA,
  alpha_shape = NA,
  alpha_rate = NA,
  mu_phi = NA,
  calendar_ages = NA,
  n_clust = min(10, length(c14_determinations)),
  use_cpp = TRUE
)
}
\arguments{
\item{c14_determinations}{A vector of observed radiocarbon determinations}

\item{c14_sigmas}{A vector of the radiocarbon determinations
uncertainties (1-sigma). Must be the same length as \code{c14_determinations}.}

\item{calibration_curve}{A dataframe which should contain at least 3 columns
entitled \code{calendar_age}, \code{c14_age} and \code{c14_sig}.
This format matches \link{intcal20}.}

\item{n_iter}{The number of MCMC iterations (optional). Default is 100.}

\item{n_thin}{How much to thin the output (optional). 1 is no thinning,
a larger number is more thinning. Default is 10. Must choose an integer more
than 1 and not too close to \code{n_iter}, since after burn-in there are
\eqn{(n_{\textrm{iter}}/n_{\textrm{thin}})/2} samples from posterior to
potentially use.}

\item{slice_width}{Parameter for slice sampling (optional). Default is
either 1000 or half the width of the c14 determinations, whichever is larger.}

\item{slice_multiplier}{Integer parameter for slice sampling (optional).
Default is 10. Limits the slice size to \code{slice_multiplier * slice_width}.}

\item{show_progress}{Whether to show a progress bar in the console during
execution. Default is \code{TRUE}.}

\item{sensible_initialisation}{Whether to use sensible start values and
adaptive prior on \eqn{\mu_{\phi}} and  (A, B).
If this is \code{TRUE} (the default), then all the remaining arguments below are
ignored.}

\item{lambda, nu1, nu2}{Hyperparameters for the prior on the means
\eqn{\phi_j} and precision \eqn{\tau_j} of each individual calendar age
cluster \eqn{j}.
\deqn{(\phi_j, \tau_j)|\mu_{\phi} \sim
\textrm{NormalGamma}(\mu_{\phi}, \lambda, \nu_1, \nu_2)} where
\eqn{\mu_{\phi}} is the overall cluster centering. Required if
\code{sensible_initialisation} is \code{FALSE}.}

\item{A, B}{Prior on \eqn{\mu_{\phi}} giving the mean and precision of the
overall centering \eqn{\mu_{\phi} \sim N(A, B^{-1})} i.e.
B small is uninformative. Required if \code{sensible_initialisation} is \code{FALSE}.}

\item{alpha_shape, alpha_rate}{Hyperparameters for the shape and rate on prior
for DP concentration, \eqn{\alpha}, determining the number of clusters we
expect to observe among our n sampled objects.
\eqn{\alpha \sim \Gamma(\eta_1, \eta_2)} where \eqn{\eta_1, \eta_2} are
the \code{alpha_shape} and \code{alpha_rate}. A small alpha means more concentrated
(i.e. few clusters) while a large alpha means not concentrated (i.e. many
clusters).  Required if \code{sensible_initialisation} is \code{FALSE}.}

\item{mu_phi}{Initial value of the overall cluster centering \eqn{\mu_{\phi}}.
Required if \code{sensible_initialisation} is \code{FALSE}.}

\item{calendar_ages}{The initial estimate for the underlying calendar ages
(optional). If supplied it must be a vector with the same length as
\code{c14_determinations}.  Required if \code{sensible_initialisation} is \code{FALSE}.}

\item{n_clust}{The initial number of clusters (optional). Must
be less than the length of \code{c14_determinations}. Default is 10 or the length
of \code{c14_determinations} if it is less than 10.}

\item{use_cpp}{DEV ONLY: Use new cpp functions (should be faster).}
}
\value{
A list with 11 items. The first 8 items contain output data, each of
which have one dimension of size \eqn{n_{\textrm{out}} =
\textrm{floor}( n_{\textrm{iter}}/n_{\textrm{thin}}) + 1}, each row storing
the result from every \eqn{n_{\textrm{thin}}}th iteration:

\describe{
\item{\code{cluster_identifiers}}{An \eqn{n_{\textrm{out}}} by
\eqn{n_{\textrm{obs}}} integer matrix. Gives the cluster allocation
- an integer between 1 and n_clust - for each observation.}
\item{\code{alpha}}{A double vector of length \eqn{n_{\textrm{out}}} giving the DP
concentration parameter.}
\item{\code{n_clust}}{An integer vector of length \eqn{n_{\textrm{out}}} giving
the number of clusters.}
\item{\code{phi}}{A list of length \eqn{n_{\textrm{out}}} each entry giving
a vector of the cluster means \eqn{\phi_j}.}
\item{\code{tau}}{A list of length \eqn{n_{\textrm{out}}} each entry giving
a vector of the cluster uncertainties \eqn{\tau_j}.}
\item{\code{weight}}{A list of length \eqn{n_{\textrm{out}}} each entry giving
the mixing weights of each cluster.}
\item{\code{calendar_ages}}{An \eqn{n_{\textrm{out}}} by \eqn{n_{\textrm{obs}}}
integer matrix. Gives the calendar age for each observation.}
\item{\code{mu_phi}}{A vector of length \eqn{n_{\textrm{out}}} giving the overall
centering \eqn{\mu_{\phi}} of the clusters.}
}
where \eqn{n_{\textrm{obs}}} is the number of radiocarbon observations i.e.
the length of \code{c14_determinations}.

The remaining items give information about input data, input parameters (or
those calculated using \code{sensible_initialisation}) and update_type

\describe{
\item{\code{update_type}}{A string that always has the value "Walker".}
\item{\code{input_data}}{a list containing the C14 data used and the name of
the calibration curve used.}
\item{\code{input_parameters}}{A list containing the values of the fixed
hyperparameters \code{lambda}, \code{nu1}, \code{nu2}, \code{A}, \code{B}, \code{alpha_shape},
\code{alpha_rate} and \code{mu_phi}, and the slice parameters \code{slice_width} and
\code{slice_multiplier}.}
}
}
\description{
This function takes as an input a set of radiocarbon determinations and
associated 1-sigma uncertainties, as well as the calibration curve which
should be used, and returns output data that can be sampled to estimate the
joint calendar age density and cluster.
}
\examples{
# Basic usage making use of sensible initialisation to set most values and
# using a saved example data set. Note iterations are kept very small here
# for a faster run time.
WalkerBivarDirichlet(kerr$c14_ages, kerr$c14_sig, intcal20, n_iter=100, n_thin=10)
}
